Added the security

+++ src/main/java/com/project/E_Commerce/dto/User/SearchResponse.java
@@ -8,5 +8,5 @@ public interface SearchResponse {
     BigDecimal getPrice();
     String getBrandName();
     String getImageUrl();
-    Double getAverageRating();
+    Integer getAverageRating();
 }
